Fixes to tests and plugin useWith.server/client spec.
Now tests work with the other plugin repos. Also fixed a few bugs introduced by
the last test suite re-write.

Plugins now come with a plugin.useWith = {server: true, browser: true}
descriptor, so the test runner can determine which plugins to apply to the
server racer versus the client racer. This also serves as documentation to the
developer, so he/she immediately knows which environemnt(s) the plugin is
inteneded to apply to.