Fundamental fix to the (num, ver) non-monotonic problem.
Before we could get into states where (num: 2, ver: 3) and (num: 3, ver: 2)
were sent down to the client. This broke monotonicity of ver in the stream of
acks to the browser client. Now, we enforce num to be monotonic with ver, on
the server-side by remembering what version to expect when looking at a
request assoc with the txn on the server after a db write, before incrementing the client
specific num for that transaction. If the expected version does not match, then
we punt on incrementing the num, until we receive and incrNum the requests that had
incremented versions putting them before the request at hand.