Fix more cases by being more loose with requirements
Especially, the case where one element is changed to undefined (through
model.del), any input will trigger blur but after the element is most
likely removed, which in turn will trigger the blur event with it's
value on the next items input. In other words, what happens is that the
next item will get the deleted items value in (some) inputs.