Fix deferred unload (see a320772a) when unfetching then synchronously subscribing
When a doc gets unfetched, #276 adds a Share Doc whenNothingPending listener to unload the doc later, instead of doing nothing, which was the cause of a memory leak introduced in #266.

However, Racer keeps reference counts itself for fetches and subscriptions, so an immediate subscribe on the same Racer doc does not affect the Share doc at all, which means the doc would still get unloaded, erroneously.

This fix instead retries the Racer _maybeUnloadDoc in that scenario, instead of doing an immediate unload. That guards against Racer getting new doc references while the Share Doc whenNotingPending is in progress.